Added Object Selector and Scale-Changer, Imported TextMeshPro
I made a very basic (and slightly buggy) object selector, and added in UI to change that object's X, Y, and Z scale (by a factor).  It uses Listeners and UnityActions to update the add/subtract the scale from the values in the GUI; mostly for demonstrations purposes.  The objected is highlighted a light-blue if it can be selected, and a darker blue if it is currently selected.
We will need a way to select an object like this if we want to use 3D models for cards, rather than have them be GUI-based!
I also imported TextMeshPro, which makes up most of the changed files, to make some GUI look a bit better (though I didn't end up using it for this demo for simplicity in programming).