"Animated" Turns & Added Mana System
I added very rudimentary "animation" (Player and Enemy move forward, then back after 0.5 seconds when they use a move).  I also added in a system to make the player's moves cost mana to use, along with adding the ability to use as many moves as the Player wants per-turn, as long as they have the required mana.  I updated the GUI to show the Player's current mana, and the cost of each move (Buff and Attack.)
The idea of Buff and Attack is that they'll be replaced with the Cards themselves later, having them be buttons is just for testing.
Maybe have a list of Card objects called "Hand" on the Player class, and call Card.Play() when that Card is selected?