Card Improvements
Made a Card1 script to test specific card functions (mana, damage) as well as a new texture (to more easily recognize the card). Unfortuneately I've run into an issue with the constructors and inhieretance ("There is no argument given that corresponds to the required formal parameter 'n' of 'Card.Card(string, int, int, int)'"), if anyone is able to look over that and see if they can catch the issue.