Made a new test scene for Movement and fixed a bug with Enemy's health
I made a new scene to test the Player and Enemy Movement; right now it just contains a Player, Enemy, and Plane with no scripts attached yet.  I also added in a Damage number for the Player on the Combat scene (This should be replicated for the Enemy, as well as for their defense).  Finally, I squashed a small bug that updated the enemy's health value improperly due to a merge conflict.