UI Changes for the SampleScene and Combat
I made the Enemy and Player's Textual health value attached to the slider - the UI looks cleaner, less cramped,, and the health value moves along with the slider now.
I also edited the UI in the SampleScene to fit on the screen better - it hadn't had the same treatment as the Combat scene's yet.