TileMapGenerator and Tile class in TileMovement scene
I created a new scene to test Tile-based movement in, as well as two new scripts (Tile and TileMapGenerator) and a new prefab (Tile).  Tile currently does nothing but hold a reference to a card (ex: a Trap card played on it), and some commented-out example code of how to call that nonexistent (yet) function.
The TileMapGenerator generates a grid of tiles, based on three variables:  TileWidth, TileLength, and TileOffset.  Width/Length act as the x & y lengths of the grid, while TileOffset is the distance -between- each tile.
I made this following a tutorial; https://www.youtube.com/watch?v=3NQnpaSNsKY
I also tried to add them to a 2D-Array for easier access to a specific tile, but ran into some trouble with References.