Added tiles to 2D array, created Wall formula
I fixed an error with the logic adding the tiles to a 2D array for easy access.  I also created an algorithm that places "Wall Units" around the outer perimeter of the grid.  The y-value of its position is the wall's y-scale divided by 2, because this places its position at the exact height so that the bottom of the Wall model is at y=0.