Lighting & Camera adjustments, prepping Card class for Tiles
I changed the Directional Light to a Point Light (sphere-like) because I think it looks better for this scene.
I also added in a short formula for adjusting the Camera and Light positions based on the size of the tile grid.
I also added a single method to the Card class, which will be what the card actually 'does' when it is used on a target.  The "base" class Card does nothing, but it's important to keep the method for polymorphism.  This method will be overridden in any children Cards.
Finally, I fixed a small error in the Wall generation formula that made it only scale 2 sides of the wall, instead of all 4.