Basic Player Tile Movement
I added 4 buttons to the Tile Movement GUI; Up/Down/Left/Right, which control in which direction the Player will move next.  On clicking, the Player will move in that direction and land on the next Tile; with bounding done correctly and a system in place to prevent spamming the button.
The buttons will hopefully be replaced by being able to select a Tile and move there; I added in generalized SelectionGO & SelectableGO scripts to help with that (and remove the old InputField-oriented code for the example test scene.)