Player can Select Tiles to Move Towards
I changed a bunch of classes here, all relating to the Tiles.  I edited the logic in TileSelectable and TileSelection so that the Move button is made interactable/uninteractable if there is a tile selected or no tile selected, and it also leaves tiles the same Cyan color if deselected (since they can still be re-selected.)  I also made it so that by default, the Tiles are not selectable - this has to be manually enabled by the PlayerClickToMove script's PaintReachableTiles function; which also paints the tiles Cyan.
I had a lot of trouble getting the "angle" in PlayerClickToMove to judge which Cardinal direction is the "closest."  I tried several guides but none worked, so I had to think it through myself!  It creates a temporary game object, gets the angle between tiles by using the LookAt() function, and then based on where the angle lies in a series of if-else statements, it moves in that direction.
I also added another button for Ending the Player's turn, which just resets how many tiles it can move.
Writing the PaintReachableTiles method was easier than I thought it'd be.  I didn't come up with a unique method, but I did come up with it on my own!