Reading Text File; Updating Deck
This is incomplete, but I'm attempting to display the information of a text file onto a UI. Also, I am attempting to see if how I'm using GenerateDeck and CardHandler are actually working by writing to a text file and seeing if the information displays. I believe that I have to change it "toString" somehow, but I ran out of time for now.