Prefabs for Enemies, GUI on each Enemy
I made prefabs for each Enemy, and also placed a GUI similar to the ones on the Card prefabs on each enemy.  This is a convenient way to position the new GUI above each model, and also keeps the code for assigning them simple; dragging and dropping the assignments inside of the prefab editor works just fine!