CardTestParticle "shoots" a Fireball at Enemies
I made a new card to test a particle propulsion system; "LerpTowardsTargets" that moves a Particle System towards a target, and then also Instantiates a sub-particle system when it reaches the target.
In this case, it sends a Fireball towards the target, and then instantiates some Smoke that rises up afterwards.
I also changed the Heal method to not require negative values - it seemed like it would cause confusion.