Changed Player pathing algorithm to AStar
The Player's pathing algorithm is now also AStar, to match the enemies'.  When a tile is Selected, the AStar path to that tile is calculated.  Then, the Tiles along that path are colored - RED means the Player does not have enough movement left to reach that tile, while GREEN means that Player will reach it when they hit the Move button.