Updated Combat Scene
Now loads the same Combat Scene as Sample Combat, but without the Player - since the Player is Global now, having another Player in the scene would cause 2 players to be stored globally... and another one added each time the Scene is loaded.  Not good!
Instead of having the Player loaded in with the Scene, the Turns class now has the responsibility of finding a Player (FindObjectOfType<Player>()) and placing them at the appropriate position & rotation.