Implemented Card Deck, Hand, Graveyard, and List of All Cards
I implemented all of the above.  Apparently Unity handles the size of objects on its own when they're Instantiated; I just had to move the cards over a bit so they weren't clipping into each other since they're all instantiated at the same time.
I implemented all 4 of the Classes fairly-close to the Graph that I drew up for them a few weeks ago.

When the Player ends their turn, new cards up to the HandSize (in the Hand class for now, should be moved to the Player; rather easily done) are drawn from the Deck class.  If the Deck does not have enough cards to fill the Hand, it will Reshuffle every card from the Graveyard class into the Deck, then resume filling the Hand.  Once a Card is played, it's removed from the Hand, the positions of all other Cards in the Hand are updated, and the played card is Discarded into the Graveyard.