Started work on Shop Scene, and Redid Card Prefabs
I started the Shop Scene; right now it generates 6 random Cards from the List of All Cards and displays them above Buy buttons.  Clicking the Buy button removes the Card and adds it to the Player's Deck.
Later, I'll implement the Leave button (Which will take the Player back to the Tile Movement Scene) and the Destroy A Card button, which will allow the Player to remove any Card from their Deck.