Added Event TurnEnded to Turns Class, and progress on the Shop Scene
There is now a listener for when both the Enemy's and Player's turns are over - this will be useful for things like Status Effects.  I made some changes to the Shop class, as well as some changes to the Deck, Hand, and Graveyard classes - They now reset when a new scene loads.