Adjusted Poison & Status Effect Classes
Apparently embedded classes do not instantiate correctly with MonoBehavior, so they need to be separate classes if they rely on the Start() and Awake() and Update() methods.  I also added some logic to use the turnsLeft variable; when it's 0, it will remove the Status Effect from the listener and destroy itself.