More work on Shop Scene
Shop scene's Destroy a Card button now creates a list of the Player's entire Deck when clicked.  It also creates a Button below each card; when clicked, these will remove that Card from the Player's Deck.
Work still needs to be done on getting the ScrollRect to scale correctly to the amount of Cards in the deck, though the basic logic of the Card Removal works correctly.