Shop Scene GUI (mostly) Done
The Shop Scene is mostly finished now.  It just needs some models in the background to look nicer, and potentially a bit of polish on the DeckGUI script.  The only issue with DeckGUI right now is that the CardHolder object does not get *centered* on the instantiated cards.  Settings its position to that center is not easy.
There also might be some small bugs with the UI not updating correctly, need to spend more time testing this.
I also changed a bit of logic within some the Defense Down status effect, to check for negative values, and lowered the damage of Poison to 10 per turn.
I also fixed some logic with the Deck handling that would occasionally double the Deck size.