Enemies face the Player, and move directly towards them.
I edited the prefabs to fix their rotations, and also altered the Turns class.  It now makes the instantiated Enemies face towards the Player when they're spawned in, and during their attack, each Enemy moves towards the Player while attacking - this can be changed to trigger their Walking animations (if applicable) later, and to last a longer time if needed.