Added "Vampiric Bite" Lifesteal Card
Created script and prefab for a lifesteal card.
Currently not in ListOfAllcards because it's giving a null pointer exception and seems unable to set the card name. It works right, I just don't want to risk any issues with pulling if it pulls in something that causes an error when run.