Treasure Scene basic logic & PassiveItem examples implemented
I implemented the Treasure Scene's basic logic; The player goes to the Scene, opens the Chest to receive a Passive Item, and then leaves back to the Tile Movement scene.  Right now, there is a good bit of skeleton code that needs to be actually implemented; like the Inventory GUI.
If the Player has already exhausted the full list of Passive Items (currently 1), then there will be nothing to get from the Chest, and the button to Open it will not appear.
As an example PassiveItem, I made the PoisonAura script.  Whenever Combat is started, it applies the poison status effect to every enemy.  To facilitate this, I added a CombatStarted eventHandler to the Turns class; the PoisonAura subscribes & reacts to that to trigger the poisoning effect.
I also fixed a small bug that could trigger in status effects; if the Enemy dies while one is still active, it is still subscribed to the TurnEnded event, and will cause a nullReference error when that Enemy's turn happens.  Now, on death, all Enemy's status effects are purged from the TurnEnded list.