Tile Scene now randomly spawns Enemies, Movement improved
I also adjusted the Tile size to accommodate for the Enemy & Player models' large size.  The Prefabs were also slightly off, position-wise; hopefully the player & enemies will be aligned with the tiles properly now.
I also added a light to the Tile scene, since spawning in the Point Lights could be a bit unreliable lighting-wise.
Finally, I added in a line of code to the Movement class that makes the object about to move face in the direction of the movement.