Status Effects now Revert their Stat Changes
Status Changes are reverted after:
1) Their turn limit runs out
2) The Enemy/Player it is on dies
or 3) Combat ends.
I updated all the relevant Status Effect cards & their base classes to use the new Status Effect methods; Revert() and ApplyEffect().  These are only necessary on cards that apply a status effect that is "on or off", meaning it only changes a value when it's put on, and when it's taken off.  (Damage over time effects, like Poison, will just be normal status effects, and calling these methods is not necessary for them.)
There is also a new method on every status effect called UpdateValues(); the first value should be the "value" on the Card, and the second should be the number of Turns this card will last.