Fixed Positions in the TileGrid & some issues with Animators
Players and Enemies would sometimes clip through the tiles; they should be fixed now, along with the Player's rotation.
Also, the Animator controllers for the PlayerKnight & Chest are changed; the Player doesn't attack twice anymore and the Chest starts closed, and then opens only when the button is clicked.  Or if there are no more items, the chest starts opened.